Updated casl ability and user service

+++ server/src/modules/user/user.service.ts
@@ -1,37 +1,69 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UserRepository } from './user.repository';
-import { FindOneOptions, FindOptionsRelations } from 'typeorm';
 import { User } from './entities/user.entity';
+import { AccessTokenPayload } from '../../common/types/access-token-payload.type';
+import { CaslAbilityFactory } from '../../core/security/casl/casl-ability.service';
+import { Action } from '../../common/enums/action.enum';
 
 @Injectable()
 export class UserService {
 
   constructor(
     private readonly userRepository: UserRepository,
+    private readonly caslAbilityFactory: CaslAbilityFactory,
   ) {}  
   async create(createUserDto: CreateUserDto) {
-    return await this.userRepository.create(createUserDto);
+    return await this.userRepository.create({
+      ...createUserDto,
+    });
   }
 
-  async findAll() {
+  async findAll(accessTokenPayload: AccessTokenPayload) {
     return await this.userRepository.findAll({});
   }
 
-  async findOne(options: FindOneOptions<User>): Promise<User | null> {    
-    return await this.userRepository.findOne(options);
+  async findOne(accessTokenPayload: AccessTokenPayload, id: number) { 
+
+    // 1. Fetch by ID first (don't filter by userId in SQL so Admins can find it)
+    const user = await this.userRepository.findOne({ where: { id }, relations: ['roles', 'roles.permissions'] });
+    if (!user) throw new NotFoundException(`User ${id} not found`);
+    
+    // 2. Check Permissions via CASL
+    const ability = this.caslAbilityFactory.createForUser(accessTokenPayload);
+
+    // Checks if user is Admin OR if user.id === token.userId
+    if (ability.cannot(Action.READ, user)) {
+      throw new ForbiddenException(`You do not have permission to access this user`); // Hide existence if no permission
+    }
+    return user;
   }
 
-  async findById(id: number): Promise<User | null> {    
+  async findById(accessTokenPayload: AccessTokenPayload, id: number): Promise<User | null> {    
     return await this.userRepository.findOne({ where: { id }, relations: ['roles', 'roles.permissions'] });
   }
 
-  async update(id: number, updateUserDto: UpdateUserDto) {
+  async update(accessTokenPayload: AccessTokenPayload, id: number, updateUserDto: UpdateUserDto) {
+    const existingUser = await this.findOne(accessTokenPayload, id);
+    const ability = this.caslAbilityFactory.createForUser(accessTokenPayload);
+
+    // Checks if user is Admin OR if user.id === token.userId
+    if (ability.cannot(Action.UPDATE, existingUser)) {
+      throw new ForbiddenException('You do not have permission to update this user');
+    }
+    Object.assign(existingUser, updateUserDto);
     return await this.userRepository.update(id, updateUserDto);
   }
 
-  async remove(id: number) {
+  async remove(accessTokenPayload: AccessTokenPayload, id: number) {
+    const existingUser = await this.findOne(accessTokenPayload, id);
+    const ability = this.caslAbilityFactory.createForUser(accessTokenPayload);
+
+    // Checks if user is Admin OR if user.id === token.userId
+    if (ability.cannot(Action.DELETE, existingUser)) {
+      throw new ForbiddenException('You do not have permission to delete this user');
+    }
     return await this.userRepository.delete(id);
   }
 }
